Modified to use the Long parser for integers, since some values can be LLOOOONNNGGGG

+++ engine/impl/ucm/src/main/java/com/armedia/caliente/engine/ucm/model/UcmAttributes.java
@@ -119,7 +119,7 @@ public final class UcmAttributes {
 					break;
 				case INT:
 					if (!StringUtils.isEmpty(v)) {
-						value = new CmfValue(Integer.valueOf(v.toString()));
+						value = new CmfValue(Long.valueOf(v.toString()));
 					} else {
 						T = CmfDataType.INTEGER;
 					}
